refactor: one test per file, drop solder paste snapshot matcher
